Allow development override of custom entity type permissions

+++ src/main/java/org/folio/fqm/service/EntityTypeService.java
@@ -386,8 +386,8 @@ public class EntityTypeService {
   }
 
   public CustomEntityType getCustomEntityTypeWithAccessCheck(UUID entityTypeId) {
-    var customET = getCustomEntityType(entityTypeId);
-    enforceCustomEntityTypeAccess(customET);
+    CustomEntityType customET = getCustomEntityType(entityTypeId);
+    permissionsService.verifyUserCanAccessCustomEntityType(customET);
     return customET;
   }
 
@@ -431,10 +431,10 @@ public class EntityTypeService {
   }
 
   public CustomEntityType updateCustomEntityType(UUID entityTypeId, CustomEntityType customEntityType) {
-    var oldET = getCustomEntityType(entityTypeId);
-    enforceCustomEntityTypeAccess(oldET);
+    CustomEntityType oldET = getCustomEntityType(entityTypeId);
+    permissionsService.verifyUserCanAccessCustomEntityType(oldET);
 
-    var updatedCustomEntityType = customEntityType.toBuilder()
+    CustomEntityType updatedCustomEntityType = customEntityType.toBuilder()
       .createdAt(oldET.getCreatedAt())
       .updatedAt(clockService.now())
       .owner(Objects.requireNonNullElse(customEntityType.getOwner(), oldET.getOwner()))
@@ -572,26 +572,16 @@ public class EntityTypeService {
     }
   }
 
-  private boolean currentUserCanAccessCustomEntityType(CustomEntityType customET) {
-    return customET.getShared() || folioExecutionContext.getUserId().equals(customET.getOwner());
-  }
-
   boolean currentUserCanAccessCustomEntityType(String entityTypeId) {
-    var customET = getCustomEntityType(UUID.fromString(entityTypeId));
-    return currentUserCanAccessCustomEntityType(customET);
+    CustomEntityType customET = getCustomEntityType(UUID.fromString(entityTypeId));
+    return permissionsService.canUserAccessCustomEntityType(customET);
   }
 
   void enforceAccessForPossibleCustomEntityType(UUID entityTypeId) {
     entityTypeRepository.getEntityTypeDefinition(entityTypeId, executionContext.getTenantId())
       .filter(et -> Boolean.TRUE.equals(et.getAdditionalProperty("isCustom")))
       .map(et -> getCustomEntityType(entityTypeId))
-      .ifPresent(this::enforceCustomEntityTypeAccess);
-  }
-
-  void enforceCustomEntityTypeAccess(CustomEntityType customET) {
-    if (!currentUserCanAccessCustomEntityType(customET)) {
-      throw new CustomEntityTypeAccessDeniedException("Entity type " + customET.getId() + " is not shared. It can only be accessed by its owner");
-    }
+      .ifPresent(permissionsService::canUserAccessCustomEntityType);
   }
 
   public void deleteCustomEntityType(UUID entityTypeId) {
@@ -600,7 +590,7 @@ public class EntityTypeService {
     if (customEntityType.getIsCustom() == null) {
       throw new EntityTypeNotFoundException(entityTypeId, "Entity type " + entityTypeId + " is not a custom entity type, so it cannot be deleted");
     }
-    enforceCustomEntityTypeAccess(customEntityType);
+    permissionsService.verifyUserCanAccessCustomEntityType(customEntityType);
     entityTypeRepository.deleteEntityType(entityTypeId);
   }
